+++ Assets/OcsVehicle/Scripts/Vehicle/DriveTrain/CylinderDrivenJoint_Physics.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 19a07a3a9960fbc408c8a19f54a60c99
+guid: 3ddf0240827b882c5a8c95810e8c06b1
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
